+++ Project.toml
@@ -3,9 +3,7 @@ uuid = "909927c2-98d5-4a67-bba9-79f03a9ad49b"
 version = "0.0.1-DEV"
 
 [deps]
-Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
 BlockArrays = "8e7c35d0-a365-5155-bbbb-fb81a777f24e"
-CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 CommonSolve = "38540f10-b2f7-11e9-35d8-d573e4eb0ff2"
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 DiffEqBase = "2b5f629d-d688-5b77-993f-72d75c75574e"
@@ -32,6 +30,11 @@ TimerOutputs = "a759f4b9-e2f1-59dc-863e-4aeb61b1ea8f"
 UnPack = "3a884ed6-31ef-47d7-9d2a-63182c4928ed"
 Unrolled = "9602ed7d-8fef-5bc8-8597-8f21381861e8"
 WriteVTK = "64499a7a-5c06-52f2-abe2-ccb03c286192"
+Adapt = "79e6a3ab-5dfb-504d-930d-738a2a938a0e"
+
+
+[weakdeps]
+CUDA = "052768ef-5323-5732-b1bb-66c8b64840ba"
 
 [extensions]
 CuThunderboltExt = "CUDA"
@@ -39,7 +42,6 @@ CuThunderboltExt = "CUDA"
 [compat]
 Aqua = "0.8"
 BlockArrays = "1"
-CUDA = "5.6.0"
 DiffEqBase = "6"
 Ferrite = "1"
 LinearSolve = "2"
