+++ src/ferrite-addons/gpu/gpugrid.jl
@@ -1,5 +1,5 @@
 # Utility which holds partial information for assembly.
-struct GPUGrid{sdim, C<:AbstractCell, T<:Real, CellDataType <: AbstractGPUVector{C}, NodeDataType <: AbstractGPUVector} <: AbstractGrid{sdim}
+struct GPUGrid{sdim, C<:Ferrite.AbstractCell, T<:Real, CellDataType <: AbstractGPUVector{C}, NodeDataType <: AbstractGPUVector} <: Ferrite.AbstractGrid{sdim}
     cells::CellDataType
     nodes::NodeDataType
     #TODO subdomain info
@@ -23,6 +23,28 @@ function _show(io::IO, mime::MIME"text/plain", grid::GPUGrid{sdim, C, T}, indent
     println(io, " cells and $(length(grid.nodes)) nodes")
 end
 
-function Adapt.adapt_structure(to, grid::SimpleMesh)
-    return adapt_structure(to, grid.grid)
+# commented out because SimpleMesh is not defined in FerriteUtils module.
+# function Adapt.adapt_structure(to, grid::SimpleMesh)
+#     return adapt_structure(to, grid.grid)
+# end
+
+
+Ferrite.get_coordinate_type(::GPUGrid{sdim, <:Any, T,<:Any,<:Any}) where {sdim, T} = Vec{sdim, T} # Node is baked into the mesh type.
+
+
+@inline Ferrite.getnodes(grid::GPUGrid, v::Ti) where {Ti<: Integer} = grid.nodes[v]
+
+
+function Ferrite.getcoordinates(grid::GPUGrid, e::Ti) where {Ti<: Integer}
+    # e is the element index.
+    CT = Ferrite.get_coordinate_type(grid)
+    cell = getcells(grid, e)
+    N = nnodes(cell)
+    x = MVector{N, CT}(undef) # local array to store the coordinates of the nodes of the cell.
+    node_ids = get_node_ids(cell)
+    for i in 1:length(x)
+        x[i] = get_node_coordinate(grid, node_ids[i])
+    end
+
+    return SVector(x...)
 end
