--- src/gpu/adapt.jl
@@ -1,32 +1,0 @@
-# TODO: this file should go to an extension later## TODO: put the adapt somewhere else ?!
-function Adapt.adapt_structure(to, element_cache::AnalyticalCoefficientElementCache)
-    cc = Adapt.adapt_structure(to, element_cache.cc)
-    nz_intervals = Adapt.adapt_structure(to, element_cache.nonzero_intervals |> cu)
-    cv = element_cache.cv
-    fv = Adapt.adapt(to, FerriteUtils.StaticInterpolationValues(cv.fun_values))
-    gm = Adapt.adapt(to, FerriteUtils.StaticInterpolationValues(cv.geo_mapping))
-    n_quadoints = cv.qr.weights |> length
-    weights = Adapt.adapt(to, ntuple(i -> cv.qr.weights[i], n_quadoints))
-    sv = FerriteUtils.StaticCellValues(fv, gm, weights)
-    return AnalyticalCoefficientElementCache(cc, nz_intervals, sv)
-end
-
-function Adapt.adapt_structure(to, coeff::AnalyticalCoefficientCache)
-    f = Adapt.adapt_structure(to, coeff.f)
-    coordinate_system_cache = Adapt.adapt_structure(to, coeff.coordinate_system_cache)
-    return AnalyticalCoefficientCache(f, coordinate_system_cache)
-end
-
-function Adapt.adapt_structure(to, cysc::CartesianCoordinateSystemCache)
-    cs = Adapt.adapt_structure(to, cysc.cs)
-    cv = Adapt.adapt_structure(to, cysc.cv)
-    return CartesianCoordinateSystemCache(cs, cv)
-end
-
-function Adapt.adapt_structure(to, cv::CellValues)
-    fv = Adapt.adapt(to, FerriteUtils.StaticInterpolationValues(cv.fun_values))
-    gm = Adapt.adapt(to, FerriteUtils.StaticInterpolationValues(cv.geo_mapping))
-    n_quadoints = cv.qr.weights |> length
-    weights = Adapt.adapt(to, ntuple(i -> cv.qr.weights[i], n_quadoints))
-    return FerriteUtils.StaticCellValues(fv, gm, weights)
-end
