+++ test/gpu/operators-test.jl
@@ -4,6 +4,9 @@ using CUDA
 using Test
 using StaticArrays
 
+
+test_ext(1.0)
+
 grid = generate_grid(Quadrilateral, (2,2))
 dh = DofHandler(grid)
 add!(dh, :u, Lagrange{RefQuadrilateral,1}())
@@ -42,4 +45,6 @@ Thunderbolt.update_operator!(cuda_op,0.0)
 
 
 @test Vector(cuda_op.op.b) ≈ linop.b
-|
+
+Thunderbolt.update_operator! |> methods
+
